+++ .env.example
@@ -31,6 +31,7 @@ EMAIL_USE_SSL=False
 AWS_COGNITO_CLIENT_ID=
 AWS_COGNITO_CLIENT_SECRET=
 AWS_COGNITO_DOMAIN=ngohub-staging.auth.eu-west-1.amazoncognito.com
+AWS_COGNITO_USER_POOL_ID=eu-west-abcdef_0123456789
 
 NGOHUB_API_HOST=api-staging.ngohub.ro
 NGOHUB_API_ACCOUNT=
