+++ backend/hub/social_adapters.py
@@ -1,14 +1,12 @@
-import mimetypes
-import requests
-import tempfile
+import logging
 
+import requests
 from allauth.core.exceptions import ImmediateHttpResponse
 from allauth.socialaccount.adapter import DefaultSocialAccountAdapter
 from allauth.socialaccount.models import SocialLogin
 from allauth.socialaccount.signals import pre_social_login, social_account_updated
 from django.conf import settings
 from django.contrib.auth.models import Group
-from django.core.files import File
 from django.dispatch import receiver
 from django.http import HttpRequest
 from django.shortcuts import redirect
@@ -16,8 +14,11 @@ from django.urls import reverse
 from django.utils.translation import gettext as _
 
 from accounts.models import User
-from hub.exceptions import DuplicateOrganizationException, MissingOrganizationException, ClosedRegistrationException
-from hub.models import Organization, NGO_GROUP, City, FeatureFlag
+from hub.exceptions import ClosedRegistrationException, DuplicateOrganizationException, MissingOrganizationException
+from hub.models import FeatureFlag, NGO_GROUP, Organization
+from hub.workers.update_organization import update_organization
+
+logger = logging.getLogger(__name__)
 
 
 class UserOrgAdapter(DefaultSocialAccountAdapter):
@@ -50,7 +51,7 @@ class UserOrgAdapter(DefaultSocialAccountAdapter):
 
 def create_blank_org(user, commit=True) -> Organization:
     """
-    Create a blank Organization (with a draft status) for an User
+    Create a blank Organization (with a draft status) for a User
     """
 
     org = Organization(user=user, accept_terms_and_conditions=True, status=Organization.STATUS.draft)
@@ -58,16 +59,6 @@ def create_blank_org(user, commit=True) -> Organization:
         org.save()
 
     return org
-
-
-def remove_signature(s3_url: str) -> str:
-    """
-    Extract the S3 file name without the URL signature and the directory path
-    """
-    if s3_url:
-        return s3_url.split("?")[0].split("/")[-1]
-    else:
-        return ""
 
 
 def update_user_org(org: Organization, token: str, *, in_auth_flow: bool = False) -> None:
@@ -112,73 +103,9 @@ def update_user_org(org: Organization, token: str, *, in_auth_flow: bool = False
             raise DuplicateOrganizationException(_("This NGO Hub organization already exists for another VotONG user."))
 
     org.ngohub_org_id = ngohub_id
-    ngohub_general = ngohub_org.get("organizationGeneral", {})
-    ngohub_legal = ngohub_org.get("organizationLegal", {})
-
-    org.county = ngohub_general.get("county", {}).get("name", "")
-    city_name = ngohub_general.get("city", {}).get("name", "")
-    try:
-        city = City.objects.get(county=org.county, city=city_name)
-    except City.DoesNotExist:
-        print("ERROR: Cannot find city '", city, "' in VotONG.")
-    else:
-        org.city = city
-
-    org.name = ngohub_general.get("name", "")
-    org.address = ngohub_general.get("address", "")
-    org.registration_number = ngohub_general.get("rafNumber", "")
-
-    # Import the organization logo
-    logo_url = ngohub_general.get("logo", "")
-    logo_filename = remove_signature(logo_url)
-    if not logo_filename and org.logo:
-        org.logo.delete()
-    elif logo_filename != org.filename_cache.get("logo", ""):
-        r = requests.get(logo_url)
-        if r.status_code != 200:
-            print("Logo file request status = %s", r.status_code)
-        else:
-            ext = mimetypes.guess_extension(r.headers["content-type"])
-            with tempfile.TemporaryFile() as fp:
-                fp.write(r.content)
-                fp.seek(0)
-                org.logo.save(f"logo{ext}", File(fp))
-            org.filename_cache["logo"] = logo_filename
-
-    # Import the organization statute
-    statute_url = ngohub_legal.get("organizationStatute", "")
-    statute_filename = remove_signature(statute_url)
-    if not statute_filename and org.statute:
-        org.statute.delete()
-    elif statute_filename != org.filename_cache.get("statute", ""):
-        r = requests.get(statute_url)
-        if r.status_code != 200:
-            print("Statute file request status = %s", r.status_code)
-        else:
-            ext = mimetypes.guess_extension(r.headers["content-type"])
-            with tempfile.TemporaryFile() as fp:
-                fp.write(r.content)
-                fp.seek(0)
-                org.statute.save(f"statute{ext}", File(fp))
-            org.filename_cache["statute"] = statute_filename
-
-    org.email = ngohub_general.get("email", "")
-    org.phone = ngohub_general.get("phone", "")
-    org.description = ngohub_general.get("description", "")
-
-    org.legal_representative_name = ngohub_legal.get("legalReprezentative", {}).get("fullName", "")
-    org.legal_representative_email = ngohub_legal.get("legalReprezentative", {}).get("email", "")
-    org.legal_representative_phone = ngohub_legal.get("legalReprezentative", {}).get("phone", "")
-
-    org.board_council = ", ".join([director.get("fullName", "") for director in ngohub_legal.get("directors", [])])
-
-    # TODO:
-    org.organization_head_name = ""
-
-    if org.status == Organization.STATUS.draft:
-        org.status = Organization.STATUS.pending
-
     org.save()
+
+    update_organization(org.id)
 
 
 @receiver(social_account_updated)
@@ -187,7 +114,7 @@ def handle_existing_login(sender: SocialLogin, **kwargs) -> None:
     Handler for the social-account-update signal, which is sent for all logins
     after the initial login.
 
-    We already have an User, but we must be sure that the User also has
+    We already have a User, but we must be sure that the User also has
     an Organization and schedule its data update from NGO Hub.
     """
