+++ backend/requirements-dev.txt
@@ -16,11 +16,12 @@ black==24.4.2
     # via -r requirements-dev.in
 blessed==1.20.0
     # via -r requirements.txt
-boto3==1.34.138
+boto3==1.34.140
     # via
     #   -r requirements.txt
     #   django-ses
-botocore==1.34.138
+    #   pycognito
+botocore==1.34.140
     # via
     #   -r requirements.txt
     #   boto3
@@ -29,7 +30,7 @@ brotlipy==0.7.0
     # via -r requirements.txt
 build==1.2.1
     # via pip-tools
-certifi==2024.6.2
+certifi==2024.7.4
     # via
     #   -r requirements.txt
     #   requests
@@ -52,7 +53,9 @@ coverage==7.5.4
 croniter==2.0.5
     # via -r requirements.txt
 cryptography==42.0.8
-    # via -r requirements.txt
+    # via
+    #   -r requirements.txt
+    #   pyjwt
 decorator==5.1.1
     # via ipython
 django==4.2.13
@@ -125,6 +128,10 @@ dnspython==2.6.1
     # via
     #   -r requirements.txt
     #   django-avatar
+envs==1.4
+    # via
+    #   -r requirements.txt
+    #   pycognito
 execnet==2.1.1
     # via pytest-xdist
 executing==2.0.1
@@ -201,6 +208,8 @@ ptyprocess==0.7.0
     # via pexpect
 pure-eval==0.2.2
     # via stack-data
+pycognito==2024.5.1
+    # via -r requirements.txt
 pycparser==2.22
     # via
     #   -r requirements.txt
@@ -210,7 +219,9 @@ pygments==2.18.0
     #   ipython
     #   pdbpp
 pyjwt==2.8.0
-    # via -r requirements.txt
+    # via
+    #   -r requirements.txt
+    #   pycognito
 pyproject-hooks==1.1.0
     # via
     #   build
@@ -243,6 +254,7 @@ pytz==2024.1
 requests==2.32.3
     # via
     #   -r requirements.txt
+    #   pycognito
     #   requests-oauthlib
 requests-oauthlib==2.0.0
     # via -r requirements.txt
