+++ backend/requirements.in
@@ -3,7 +3,7 @@ django~=4.2.13
 django-avatar~=8.0.0
 django-environ~=0.11.2
 django-crispy-forms~=2.1.0
-django-allauth[socialaccount]~=0.63.3 
+django-allauth[socialaccount]~=0.63.3
 
 # job scheduler
 django-q2~=1.6.2
@@ -32,6 +32,8 @@ brotlipy~=0.7.0
 # misc
 requests~=2.32.3
 django-ses~=4.1.0
+
+pycognito~=2024.5.1
 
 # database
 psycopg2-binary~=2.9.9
