+++ backend/requirements.txt
@@ -8,17 +8,18 @@ asgiref==3.8.1
     # via django
 blessed==1.20.0
     # via -r requirements.in
-boto3==1.34.138
+boto3==1.34.140
     # via
     #   -r requirements.in
     #   django-ses
-botocore==1.34.138
+    #   pycognito
+botocore==1.34.140
     # via
     #   boto3
     #   s3transfer
 brotlipy==0.7.0
     # via -r requirements.in
-certifi==2024.6.2
+certifi==2024.7.4
     # via
     #   requests
     #   sentry-sdk
@@ -89,6 +90,8 @@ django-storages==1.14.3
     # via -r requirements.in
 dnspython==2.6.1
     # via django-avatar
+envs==1.4
+    # via pycognito
 gevent==24.2.1
     # via -r requirements.in
 greenlet==3.0.3
@@ -113,10 +116,14 @@ psutil==5.9.8
     # via -r requirements.in
 psycopg2-binary==2.9.9
     # via -r requirements.in
+pycognito==2024.5.1
+    # via -r requirements.in
 pycparser==2.22
     # via cffi
 pyjwt==2.8.0
-    # via django-allauth
+    # via
+    #   django-allauth
+    #   pycognito
 python-dateutil==2.9.0.post0
     # via
     #   botocore
@@ -129,6 +136,7 @@ requests==2.32.3
     # via
     #   -r requirements.in
     #   django-allauth
+    #   pycognito
     #   requests-oauthlib
 requests-oauthlib==2.0.0
     # via django-allauth
